104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,17 +7,17 @@
 int main(void)
 {
 	int n = 98;
-	unsigned long long fib1 = 1, fib2 = 2;
+	long fib1 = 1, fib2 = 2;
 
 	int i;
 
-	printf("%llu, %llu", fib1, fib2);
+	printf("%ld, %ld", fib1, fib2);
 	i = 3;
 	for (; i <= n ; i++)
 	{
-		unsigned long long nextFib = fib1 + fib2;
+		long nextFib = fib1 + fib2;
 
-		printf(", %llu", nextFib);
+		printf(", %ld", nextFib);
 
 		fib1 = fib2;
 		fib2 = nextFib;
